added buy gateway to be able to use it for buying from browse and search. Prevents duplicate code
